requirements.txt
django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-filter==24.2
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
django-celery-beat==2.6.0
redis==5.0.7
psycopg2-binary==2.9.9
asgiref==3.8.1
pytest==8.2.2
pytest-django==4.8.0
